Extract more responsibilities from simulator
Moved expectations, relocations, IO handling and more to Run
