+++ src/Simulator/SuperH4/Operations/ControlFlowOperation.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Lhsazevedo\Sh4ObjTest\Simulator\SuperH4\Operations;
+
+readonly class ControlFlowOperation extends AbstractOperation
+{
+}
