+++ src/Simulator/SuperH4/Operations/GenericOperation.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Lhsazevedo\Sh4ObjTest\Simulator\SuperH4\Operations;
+
+readonly class GenericOperation extends AbstractOperation
+{
+
+}
